Add vault lock via API
