+++ source/main/services/browser/api.ts
@@ -4,14 +4,12 @@ import { VERSION } from "../../library/build";
 import { handleAuthPing, processAuthRequest, processAuthResponse } from "./controllers/auth";
 import { searchEntries } from "./controllers/entries";
 import { getAllOTPs } from "./controllers/otp";
-import { getVaults } from "./controllers/vaults";
-import { promptVaultUnlock } from "./controllers/unlock";
+import { getVaults, promptVaultLock, promptVaultUnlock } from "./controllers/vaults";
 import { handleError } from "./error";
 import { requireBrowserToken } from "./middleware";
 
 export function buildApplication(): express.Application {
     const app = express();
-    // app.enable("trust proxy");
     app.disable("x-powered-by");
     app.use(express.json());
     app.use((req: Request, res: Response, next: NextFunction) => {
@@ -31,7 +29,7 @@ function createRoutes(app: express.Application): void {
     router.get("/entries", requireBrowserToken, searchEntries);
     router.get("/otps", requireBrowserToken, getAllOTPs);
     router.get("/vaults", requireBrowserToken, getVaults);
-    router.post("/vaults/:id/lock", requireBrowserToken, () => {});
+    router.post("/vaults/:id/lock", requireBrowserToken, promptVaultLock);
     router.post("/vaults/:id/unlock", requireBrowserToken, promptVaultUnlock);
     app.use("/v1", router);
 }
