+++ source/main/types.ts
@@ -5,7 +5,8 @@ export * from "../shared/types";
 export enum BrowserAPIErrorType {
     AuthMismatch = "err/auth/mismatch",
     NoAPIKey = "err/key/missing",
-    NoAuthorization = "err/auth/missing"
+    NoAuthorization = "err/auth/missing",
+    VaultInvalidState = "err/vault/invalidState"
 }
 
 export interface OTP {
